refactor(booking): refine booking pages structure and update checkout routes
Rename routes from /booking-destination and /booking-package to /checkout-destination and /checkout-package
Rename page.jsx files to BookingDestinationList and BookingPackageList components
Prevent bookings with past dates (only allow today and future dates)
Improve clarity and maintainability across booking-related components
